Major rewrite to properly support windows
Due to the problem of keeping the number and paste settings in synch,
even during manual (non-plugin) setting of them, paste has been
effectively lowered to a window level setting rather than global. This
means it's possible to have a split window with paste mode in one and
not in the other.